Login.js
The error "User not found" was not implemented so I implemented it.
I added a feature where you can press enter instead of pressing the "Enter the wild!" button for faster and more effective way to log in, I also plan on implementing the same  feature for register, but for now login will be just fine.
TODO: I think there is nothing to do for the login page for now, only thing that comes up in my mind is the ability to login with your Google account, otherwise the login.js is done. If there are any problems they will pop up in the notion to do list.